Add LongMessageOrigin and extract public api IMirai.downloadLongMessage
